chore: remove redundant kafka snap to snap.py
